fix(suite-native): Resolve circular dependency with device ConfirmOnTrezorWrapperProps
