+++ scripts/circular-dependencies/madge-snapshot.txt
@@ -106,5 +106,4 @@ suite-common/wallet-utils/src/accountUtils.ts > suite-common/wallet-utils/src/st
 suite-common/wallet-utils/src/accountUtils.ts > suite-common/wallet-utils/src/transactionUtils.ts
 suite-common/wallet-utils/src/baseCurrency.ts > suite-common/wallet-utils/src/fiatConverterUtils.ts
 suite-common/wallet-utils/src/index.ts > suite-common/wallet-utils/src/accountUtils.ts > suite-common/suite-types/src/index.ts > suite-common/suite-types/src/modal.ts > suite-common/wallet-types/src/index.ts > suite-common/wallet-types/src/solanaStaking.ts > suite-common/staking-solana/src/index.ts > suite-common/staking-solana/src/utils/index.ts > suite-common/staking-solana/src/utils/stakingUtils.ts
-suite-native/device/src/components/ConfirmOnTrezor/ConfirmOnTrezorWrapper.tsx > suite-native/device/src/components/ConfirmOnTrezor/ConfirmOnTrezorContent.tsx
 suite-native/device/src/index.ts > suite-native/device/src/hooks/useDetectDeviceError.tsx > suite-native/device/src/selectors.ts > suite-native/firmware/src/index.ts > suite-native/firmware/src/components/FirmwareInstallationScreenContent.tsx
